requirements.txt
selenium==4.22.0
pyyaml==6.0.1
pytest==8.2.2